chore: break cyclic dev dependencies
CVLR is broken into sub-creates with a top-level crate cvlr intended
to be imported to use all of them.

This creates an issue for tests in sub-crates because they have to
depend on cvlr as a dev-dependency. However, crates.io does not allow
cyclic dependencies, even as a dev.

A solution is to ensure that dev-dependencies are specified only
through `path` (and not `version`). These are ignored by creates.io.

Reference: https://github.com/rust-lang/cargo/issues/4242