feat: support for logging scopes
`CvlrLogger::log_scope_start` and `CvrlLogger::log_scope_end` mark the
beginning and the end of the scope, respectively. Visually, the scope
is identified by nesting or indent.

The main application for scope is to log structs, so that the fields
of the logged struct are under the scope of the struct.

Logging structs is not an exclusive use, and other uses are possible.

Every started scope must end. Failure to do so is an undefined behavior.