chore: inline cmp methods in NativeIntU64
Some methods in core traits take arguments by reference.
This impedes optimization when the arguments are numeric constants.

Normally, rustc inlines all constants. However, if they have to be
passed by reference, this is not possible. By forcing the inlining of
these methods, we can improve the performance of the library
(and, more importantly, simplify the code for the Certora Prover).